Move `barrier<thread_scope_block>` to its own file
